change Stop exception handling

+++ scooterMS/src/main/java/tpe/scooterMS/service/StopService.java
@@ -9,6 +9,7 @@ import org.springframework.transaction.annotation.Transactional;
 import jakarta.validation.Valid;
 import tpe.scooterMS.DTO.DTOStopRequest;
 import tpe.scooterMS.DTO.DTOStopResponse;
+import tpe.scooterMS.exception.NotFoundException;
 import tpe.scooterMS.model.Stop;
 import tpe.scooterMS.repository.StopRepository;
 
@@ -20,49 +21,37 @@ public class StopService {
 	private StopRepository repository;
 	
 	@Transactional ( readOnly = true )
-	public List<DTOStopResponse> findAll() throws Exception {
-		try {
-			return repository.findAll().stream().map( DTOStopResponse::new ).toList();
-		} catch (Exception e) {
-			throw new Exception(e.getMessage());
-		}
+	public List<DTOStopResponse> findAll() {
+		return repository.findAll().stream().map( DTOStopResponse::new ).toList();
 	}
 	
 	@Transactional
-	public DTOStopResponse save(@Valid DTOStopRequest request) throws Exception {
-		try {
-			Stop stop = new Stop(request.getId(), request.getLatitude(), request.getLongitude());
-			stop = repository.save(stop);
-			return new DTOStopResponse(stop);
-		} catch (Exception e) {
-			throw new Exception(e.getMessage());
-		}
+	public DTOStopResponse save(@Valid DTOStopRequest request) {
+		Stop stop = new Stop(request.getId(), request.getLatitude(), request.getLongitude());
+		stop = repository.save(stop);
+		return new DTOStopResponse(stop);
 	}
 	
 	@Transactional ( readOnly = true )
-	public DTOStopResponse getStopById(long id) throws Exception {
-		try {
+	public DTOStopResponse getStopById(long id) throws NotFoundException {
+		if (repository.existsById(id)) {
 			return new DTOStopResponse(repository.getStopById(id));
-		} catch (Exception e) {
-			throw new Exception(e.getMessage());
+		} else {
+			throw new NotFoundException("Stop", id);
 		}
 	}
 	
 	@Transactional ( readOnly = true )
-	public List<DTOStopResponse> getStopsBySimpleOrdering() throws Exception {
-		try {
-			return repository.getStopsBySimpleOrdering().stream().map( DTOStopResponse::new ).toList();
-		} catch (Exception e) {
-			throw new Exception(e.getMessage());
-		}
+	public List<DTOStopResponse> getStopsBySimpleOrdering() {
+		return repository.getStopsBySimpleOrdering().stream().map( DTOStopResponse::new ).toList();
 	}
 	
 	@Transactional
-	public void deleteStop(long id) throws Exception {
-		try {
-	          repository.deleteStopById(id);
-	    } catch (Exception e) {
-	        throw new Exception(e.getMessage());
-	    }
+	public void deleteStop(long id) throws NotFoundException {
+		if (repository.existsById(id)) {
+			repository.deleteStopById(id);
+		} else {
+			throw new NotFoundException("Stop", id);
+		}
 	}
 }
